updated the animations
