+++ interactivexaiweb/package.json
@@ -9,34 +9,40 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@radix-ui/react-avatar": "^1.0.4",
-    "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-progress": "^1.0.3",
-    "@radix-ui/react-radio-group": "^1.1.3",
-    "@radix-ui/react-select": "^2.0.0",
-    "@radix-ui/react-separator": "^1.0.3",
-    "@radix-ui/react-slider": "^1.1.2",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-toast": "^1.1.5",
+    "react": "^18",
+    "react-dom": "^18",
+    "next": "14.2.5",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "typescript": "^5",
+    "tailwindcss": "^3.4.1",
     "autoprefixer": "^10.0.1",
+    "postcss": "^8",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
-    "lucide-react": "^0.263.1",
-    "next": "14.2.5",
-    "postcss": "^8",
-    "react": "^18",
-    "react-dom": "^18",
-    "recharts": "^2.8.0",
     "tailwind-merge": "^2.0.0",
-    "tailwindcss": "^3.4.1",
+    "lucide-react": "^0.263.1",
+    "@radix-ui/react-slot": "^1.0.2",
+    "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
+    "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-separator": "^1.0.3",
+    "@radix-ui/react-toast": "^1.1.5",
+    "@radix-ui/react-avatar": "^1.0.4",
+    "@radix-ui/react-radio-group": "^1.1.3",
+    "@radix-ui/react-slider": "^1.1.2",
+    "@radix-ui/react-progress": "^1.0.3",
+    "@radix-ui/react-tabs": "^1.0.4",
+    "@radix-ui/react-accordion": "^1.1.2",
+    "recharts": "^2.8.0",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "^5"
+    "react-dropzone": "^14.2.3",
+    "react-icons": "^4.12.0",
+    "react-spinners": "^0.13.8",
+    "sonner": "^1.4.3",
+    "next-themes": "^0.2.1"
   },
   "devDependencies": {
     "eslint": "^8",
